Started refactoring search tests.
